+++ app/lib/bloc/document_state.dart
