+++ app/lib/dialogs/background.dart
@@ -1,6 +1,6 @@
 import 'package:butterfly/api/open_help.dart';
 import 'package:butterfly/models/backgrounds/box.dart';
-import 'package:butterfly/pad/bloc/document_bloc.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
