+++ app/lib/dialogs/color_pick.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:butterfly/api/open_help.dart';
 import 'package:butterfly/models/palette.dart';
-import 'package:butterfly/pad/bloc/document_bloc.dart';
-import 'package:butterfly/pad/dialogs/open.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
+import 'package:butterfly/dialogs/open.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
