+++ app/lib/dialogs/elements/paint.dart
@@ -1,5 +1,5 @@
 import 'package:butterfly/models/elements/pen.dart';
-import 'package:butterfly/pad/bloc/document_bloc.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
