+++ app/lib/dialogs/settings.dart
@@ -1,4 +1,4 @@
-import 'package:butterfly/pad/bloc/document_bloc.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
