+++ app/lib/tool/view.dart
@@ -1,8 +1,8 @@
-import 'package:butterfly/pad/bloc/document_bloc.dart';
-import 'package:butterfly/pad/cubits/transform.dart';
-import 'package:butterfly/pad/dialogs/background.dart';
-import 'package:butterfly/pad/dialogs/color_pick.dart';
-import 'package:butterfly/pad/dialogs/waypoints.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
+import 'package:butterfly/cubits/transform.dart';
+import 'package:butterfly/dialogs/background.dart';
+import 'package:butterfly/dialogs/color_pick.dart';
+import 'package:butterfly/dialogs/waypoints.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
