+++ app/lib/views/toolbar.dart
@@ -1,6 +1,6 @@
-import 'package:butterfly/pad/bloc/document_bloc.dart';
-import 'package:butterfly/pad/tool/edit.dart';
-import 'package:butterfly/pad/tool/view.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
+import 'package:butterfly/tool/edit.dart';
+import 'package:butterfly/tool/view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
