+++ app/lib/views/view.dart
@@ -8,12 +8,12 @@ import 'package:butterfly/models/elements/image.dart';
 import 'package:butterfly/models/elements/label.dart';
 import 'package:butterfly/models/elements/pen.dart';
 import 'package:butterfly/models/elements/path.dart';
-import 'package:butterfly/pad/bloc/document_bloc.dart';
-import 'package:butterfly/pad/cubits/transform.dart';
-import 'package:butterfly/pad/dialogs/elements/eraser.dart';
-import 'package:butterfly/pad/dialogs/elements/image.dart';
-import 'package:butterfly/pad/dialogs/elements/label.dart';
-import 'package:butterfly/pad/dialogs/elements/paint.dart';
+import 'package:butterfly/bloc/document_bloc.dart';
+import 'package:butterfly/cubits/transform.dart';
+import 'package:butterfly/dialogs/elements/eraser.dart';
+import 'package:butterfly/dialogs/elements/image.dart';
+import 'package:butterfly/dialogs/elements/label.dart';
+import 'package:butterfly/dialogs/elements/paint.dart';
 import 'package:butterfly/models/painters/image.dart';
 import 'package:butterfly/models/painters/painter.dart';
 import 'package:butterfly/models/painters/path_eraser.dart';
